+++ blobs/micromamba_darwin.go
@@ -1,0 +1,10 @@
+package blobs
+
+import (
+	"embed"
+)
+
+//go:embed assets/micromamba.darwin_amd64.gz
+var micromamba embed.FS
+
+var micromambaName = "assets/micromamba.darwin_amd64.gz"